fix(warnings): Achieve zero-warning build across 16 source files\n\nFixed all MSVC W4 warnings:\n- Replace ::tolower with safe lambda in std::transform (13 occurrences, 7 files)\n- Fix 32-bit shift to 64-bit (1u -> 1ull) in ZstdFrameDecoder.h\n- Fix wchar_t -> char narrowing in CodecModuleSpecs.h\n- Replace deprecated GetVersionExW with RtlGetVersion in CrashIntelligence.h\n- Add [[maybe_unused]] for intentionally unused params (5 files)\n- Fix std::fill int -> uint8_t cast in BufferPoolAllocator.h\n\nBuild: 0 errors, 0 warnings. Tests: 2171/2171 pass.

+++ Engine/Memory/BufferPoolAllocator.h
@@ -68,7 +68,7 @@ struct PooledBuffer {
     bool fromPool = false;
 
     bool IsValid() const { return data != nullptr && capacity > 0; }
-    void Clear() { if (data) std::fill(data, data + capacity, 0); }
+    void Clear() { if (data) std::fill(data, data + capacity, static_cast<uint8_t>(0)); }
 };
 
 // ─── Per-class slab pool ──────────────────────────────────────────
